fixed error when no gifs
