+++ utlis/__init__.py
@@ -1,3 +1,1 @@
 from .from_admin import from_admin
-from .get_chat import get_chat
-from .get_gif import get_gif
