--- utlis/get_chat.py
@@ -1,6 +1,0 @@
-from models.chats import Chats
-
-
-async def get_chat(chat_id) -> Chats:
-    chat = await Chats.get_or_create(id=chat_id)
-    return chat[0]
